Import Collector from kalibro-processor anColld refactor Runner
Runner now just takes the desired metrics and the code path and chooses
the appropriate collectors to call.

Collector methods are now instance methods, which makes it more generic
for clients to decide how their implementations are done. It makes
Runner::BaseMetricRunner unecessary, and most of the code in
Runner::Main was moved to runner.

Signed off by: Heitor Reis <marcheing@gmail.com>