Make Kolekti register and deregister collectors by its class name
For that, create a DummyClass and AnotherDummyClass factories so that,
on tests, the instance of is not passed as a parameter to the
collector but its class name.

Kolekti#unregister_collector was refactored to keep the unit tests
working for now. Because we are unregistering collectors by its class
name, it is not possible to know which instance the user means to
unregister. But that it is not a problem for later we will be able to
register only one instance of a collector.